requirements.txt
Django==2.2.3
django-admin==1.3.2
django-excel-response2==2.0.8
django-six==1.0.4
djangorestframework==3.10.1
pymongo==3.8.0
python-dateutil==2.8.0
requests==2.22.0
